+++ RELEASE_NOTES.md
@@ -1,3 +1,5 @@
+- Fixed range constructors for empty and null arrays and sequences
+
 ### 5.3.10
 - [IO] Added SymMap constructor overload